+++ RELEASENOTES.md
@@ -11,6 +11,7 @@
     *   Add workaround for exception thrown due to `MediaMuxer` not supporting
         negative presentation timestamps before API 30.
     *   Relax trim optimization H.264 level checks.
+    *   Add support for changing between SDR and HDR input media in a sequence.
 *   Track Selection:
 *   Extractors:
 *   Audio: